backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
bcrypt
python-jose[cryptography] # Ganti PyJWT untuk keamanan yang lebih baik
pandas
openpyxl
SQLAlchemy
PyMySQL
passlib[bcrypt]
python-multipart 